pyproject.toml
[tool.poetry]
name = "django-auth-app"
version = "0.1.0"
description = ""
authors = ["Mitch <liao02x@gmail.com>"]
readme = "README.md"
packages = [{include = "django_post_app"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.31.0"
django-filter = "^23.2"
django-q = "^1.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
